requirements.txt
streamlit>=1.28.1
langchain>=0.1.0,<0.2.0
langchain-openai>=0.0.2,<0.1.0
langchain-community>=0.0.10,<0.1.0
langchain-chroma>=0.1.0,<0.2.0
python-dotenv>=1.0.0
wikipedia>=1.4.0
chromadb>=0.4.18,<0.5.0
numpy>=1.24.3,<1.25.0
pandas>=2.0.3,<2.1.0
langchainhub